Add files via upload
This is the first demo we've sent to Natty and Chris (Yale Med School Students) showing them some of our Animation Capabilities. Specifically, we wanted to show that we could animate separate parts, the camera, the syringe, and also change the texture of the human model. Waiting back to hear from them now!